refact: 컴포넌트 구조 변경 및 useComment hook 정의 (#4)
